Separate HTML attributes from adjusted XML attrs
* Use a table in tokenizer for attributes
* Pass a separate xmlAttrs seq for adjusted XML attributes

This way, deduplication is easier, and atoms only need a hash function
instead of a cmp function. Also, the number of copies is reduced in
the most common case (which is elements with html-only attributes).

Since the HTML standard does not specify the ordering of attributes,
it seems wiser to just output the table we use for de-duplication and
leave any further sorting to the consumer than to implement it inside
the library.