minidom: UTF8 validate strings in strToAtom
If we UTF8 validate, then let's go all the way.

Doing this in atomToStr would be horrible, because then different
atoms could potentially be represented by the same string.

To avoid this problem, we validate in strToAtom instead. Since
htmltokenizer/htmlparser does not keep record of strings used to
produce atoms, and never compares the string representation of atoms,
this is safe to do. (In fact, no atom stringifier is required or used
by the parser at all.)