Get rid of radixtree
entityMap was a global variable initialized at runtime; plus it wasn't
even efficient because radixtree used a heap-allocated tree, Nim
strings as keys, a closure for searching, etc.

We could have just brute forced the problem using a hash table, but
a) we would need to store hashes + dummy entries for prefixes; that's
a waste of bytes, and
b) with std tables at least, we would need to re-hash for every char
consumed.

Instead, we just use what used to be called entityTable (renamed to
entityMap), and pseudo-linear search it:
* Use a jump table to find the starting entry's index
* Go back in entityMap until either there are no more entries with the
  desired character or a matching entry found.