Unify decoder and validator APIs
The validator/decoder split was an abomination, and having to buffer
your input queue for UTF-8 inputs was kind of annoying too.

Now we have an extra instruction that also instructs the reader to use
data from its input queue. Currently this is only used for UTF-8, but we
may extend it to other ASCII-compatible charsets too in the future.