gmifetch: rewrite in Nim
This finally makes it possible to use socks5 for Gemini.

Also slightly refactored the config, to make it easier to pass on the
config dir.

By the way, the known_hosts file is now stored in the config dir too.
The adapter will try to move it to there from the old location.