pager, select: add right click menu, fix some bugs
It's *not* a context menu: items are fixed, and currently not even
customizable. Former is a feature, latter is a bug.

Also, select now has improved mouse support; its behavior is a
combination of the w3m menu (for btn1) and GTK Firefox context menu
(for btn3).

Also, fix some bugs in select:

* lines with double width chars are handled better (not perfectly,
  but at least the grid isn't completely mangled anymore)
* non-multiple select now highlights the currently selected option
* small selects at the bottom of the screen are handled correctly
* selects at the right edge of the screen are handled correctly
* select multiple no longer discards selected items on cursorLeft