pager: improve rewrite-url semantics
Previously, it just changed the URL before loading the site; now it's
an actual redirect.

Technically, the previous behavior was more flexible, because it let you
apply siteconf rules exclusively for sites where you redirected from.
Practically, this was not very useful, and probably unexpected for
anybody trying to use the feature.

This also fixes a bug where the loader filter would be set for the
original page, so you couldn't switch from https to http, etc.