container: fix crash on isearch between redirects
Sadly, pushCursorPos may not be paired with a popCursorPos in case the
container in question is replaced during the isearch. The easiest way to
reproduce this is:

* start request to a page that redirects
* start isearch
* redirection happens, now the old container is gone
* type something

Then, popCursorPos would try to pop the cursor position from the new
container.

An alternative (and better) solution would be to add a weak ref to the
container as the line edit data. Sadly, we don't have weak refs, and I
don't want to hack them in with finalizers.

(But maybe I should. Hmm.)