layout: fix initial flex item width calculation
Before, the initial layout of a flex item would have inherited the
parent's sizing constraint. This almost worked, except when a descendant
of a flex item with an unspecified width would resolve its percentage
against this incorrectly inherited width - per standard, this must
resolve to auto.

Also, the shrink case was wrong, because it did not scale the unit to
the respective widths; the standard mandates this as well. Hopefully I
got it right this time.

Finally, this fixes positioned inline container blocks not being set
as the absolute container.