buffer: fix hover text not showing up with pseudo elements
This was a bit annoying because it triggered the popup warning on
cursorNextLink -> click with form buttons (as the square brackets around
[button] are implemented with ::before/::after.)

Notably, this also removes the visibility check from isClickable.
I suspect there was a good reason why I added it, but I can't remember
why, and it's incompatible with what desktop browsers do on:

<a href="https://asdf.com" style="visibility: hidden">
<div style="visibility: visible">
test
</div>
</div>

So if it's ever added back, then this case should be respected too.