layout: temporary hack to fix xminwidth + flex
ugly, but better than the status quo.

the core issue is that width, min-width should clamp xminwidth too, but
it doesn't. this hadn't been an issue until recently, but since I made
<pre> stretch out xminwidth too it had the unfortunate result that you
now have to scroll on pages that nest pre in a flex.

this still isn't a proper fix, in particular it won't work if min-width
is set to 0, but on the sites I tried it happens to work anyway.