loader: fix heisenbug
Unconditionally registering output handles on resume could result in a
failed assertion on double-register.

The bug would appear like:
resume (register) -> handleRead -> pushBuffer -> currentBuffer is nil ->
register again

handleWrite had a very high likelihood of occurring between resume and
handleRead, and that hid the bug by immediately unregistering the handle
after resume. In fact, I haven't been able to reproduce the bug at all
today, and only found it after poring over the source...

Fix this by not registering output handles that are empty (except
if the istream is already finished, in which case it will just be
unregistered).