pager: hide kitty images when menu is open
So I thought this was just a simple z-ordering issue, but silly me,
it's never simple with images.  In this case it turns out Kitty can't
really do z-ordering with text the way Sixel can - in short, you must
pick if the image is below text, or above text, but never both.

I imagine you could also get something to work with z=-1 and stretched
1-pixel colored images or some similarly horrifying hack.  It seems very
annoying to code and maintain, so I won't.

(In some way this is impressive, because Sixel z-ordering sucks too.
Somehow we got into a situation where both viable image display
protocols are incapable of expressing some useful ways of image
layering, of course in a mutually incompatible way.)