response: simplify text()
It gets copied once anyway, so just use a blob.

I guess I could skip the copy with some effort, but the 4 lines
implementation is too attractive :P
This is still an improvement, as it doesn't needlessly zero-fill the
buffer on realloc.

(I've also removed the final realloc from blob, as it seemed quite
pointless.  Using Content-Length could help... except it doesn't,
because it refers to the encoded length. Ugh.)