term: allow ANSI fgcolor + bgcolor
Not quite sure why I had assumed that this is broken. At least on XTerm,
the previous behavior definitely was, e.g. with a white-ish background
and white foreground it would end up correcting the contrast to purple
even with black CSS color (thus breaking the "no fgcolor with bgcolor"
assumption anyway.)