Add chabookmark script
Finally it's done.  It's basically w3mbookmark, but using Markdown
instead of HTML and in POSIX shell instead of C.

As a bonus, it can also (sort of) import w3mbookmark's output.  Well,
at least it worked on my bookmark file, but there is a known issue with
bracket escaping...  if it goes wrong, it's simple enough to edit it
manually :P