match: optimize dependency tracking
This is a bit tricky, but it seems to work.  It optimizes selectors in
the line of "div :hover" (note the space.)

Previously such selectors would add a hover dependency to pretty much
every element, and trigger a re-style for all elements that changed
their hover status.  After this patch, when :hover and friends would
return false, they first try to match the element *without* pseudo
selectors, and only add their dependencies if the element would match
like that.

(Notably, it only does this for when :hover is false.  Probably it would
help somewhat if we checked for the opposite with true too, but I'm not
sure how much.

For now, I'll keep it like this, and maybe try to further optimize it
later.)