layout, render: implement overflow property
Overflow pretty much requires scrollbars, but those wouldn't work in
dump mode, plus of course they would be a pain to implement.  So as a
simple alternative:

* overflow: hidden, clip works as per spec.
* overflow: auto, overlay, scroll invert the intrinsic minimum size
  clamping logic instead of adding a scrollbar.

What this concretely means, is that this

<pre style="overflow: scroll; height: 1em">
test
test
test
</pre>

will, instead of creating a scroll container, just override the
specified height.

This hack works surprisingly well, because CSS pretty much requires
setting height on scroll containers, so authors aren't incentivized to
set height on the parent container too (because the contents are already
sized appropriately).

One issue left is how to deal with overflow: hidden ancestors.  For now,
I've made it so that it can spill for overflow-x, and always clips
on overflow-y, because it's much less likely to bleed into other
text horizontally than vertically.  But there is definitely room
for improvement, e.g. we could track space requested by scrolling
children and expand parent boxes based on that.