img: add nanosvg
Only for SVG in image tags for now.  Plus it comes with NanoSVG's
limitations, i.e. no text.  Still, better than nothing.

I've gone through open tickets and PRs at upstream to check for known
security issues.  As a recurring theme I found that parseTransform
fails to check the result of parseTransformArgs, so I fixed and
refactored that part.  (Probably not a security issue, but still UB.)

Also, I've added a fix for an OOB read in parseColorRGB.

Future directions:

* replace nanosvgrast with our own rasterizer in canvas (will have to be
  extended somewhat...)
* add text rendering (with unifont, which canvas already includes)