Misc character display fixes
* fix luwrap comparison function
* strip high unicode controls in term/pager
* use wcwidth in dirlist2html
* remove combining category from strwidth
* remove unused "disallowed" field from lineedit

My initial goal with switching to wcwidth in dirlist2html was just to
get rid of the outdated combining map in charwidth_gen.  Then I realized
that layout will normalize those out anyway, so we don't actually have
to deal with them anywhere.

Also, I found a few bugs in the process; high unicode control chars not
being stripped out was a particularly bad one, as it allows for pages to
mess up the terminal output. (Easiest way to replicate: just open a
random binary file without this patch.)