config: fix overriding substring bindings
It is unfortunately a breaking change, but the existing behavior was
practically broken anyway.

The only issue I can think of is that keybindings like `yy' now override
`y', which previously worked because y (regrettably) calls feedNext. But
those should be added to the default config anyway.

(If a user really desires the old behavior, they can just re-bind y to
cmd.buffer.copySelection after their own custom bindings. Config values
are now stored in an ordered table, so this should work.)

I've left the undocumented JS setter as it is, because the old behavior
seems more useful there. (Perhaps there should be a separate function
that behaves like toml?)