javascript: remove undocumented & broken defineConsts
I have a mostly working version in Chawan (in src/html/script.nim),
but it still has to work around the fact that we define prototypes
differently than mainstream browsers.  So I won't try to fix it here if
I'll have to change what it does *yet again* in the future.