layout: ignore general category Me, Mn
I'm not sure if this is the best way.  On one hand, it probably breaks
text in some languages that cannot be represented in normal form.
On the other hand, some terminals may crash on such strange sequences
(or worse).

I'll just skip them for now, because it's marginally faster than
returning 0 from width.  We'll see if somebody complains.