loader: add proper HTTP auth handling
Until now, we just stuffed it into the URL, which was somewhat
problematic.

Mainstream browsers like to hide the username from the user, but I've
decided to follow w3m: buffers do not receive auth info, but the pager
itself displays the username.

As for the origin: I wanted to use the regular origin, but that does not
work with any custom URL.  So instead of changing the regular origin
function, I've added another.