Try to set close-on-exec, misc cleanup
FD_CLOEXEC should hopefully get rid of bugs where buffers outlived the
main process because of some stray child process keeping the canary
pipe alive.

It's not perfect because of the runMailcapWriteFile/runMailcapReadFile
double-forks.  Ideally they should be replaced with an implementation
that tracks temporary files in the main process.