dynstream: remove safeClose, add moveFd
safeClose was originally added to prevent this bug:

let fd = doSomething() # allocates fd on 0, e.g. by opening a socket
let ps = newPosixStream(fd)
...
discard dup2(ps.fd, STDIN_FILENO)
ps.sclose() # stdin is now closed, despite the opposite intention.

With safeClose called on fds that could be stdin, the goal was that
stdin/stdout/stderr would never be allocated as a different file,
but it was still error-prone.  Enter moveFd:

ps.moveFd(STDIN_FILENO)

If ps is already stdin, this does nothing.  If not, it dup2's ps to
stdin, closes ps.fd, and sets it to stdin.