buffer: remove server socket
Now we just pass through a socket created in pager.
This removes the need for a socket directory, and strengthens the buffer
sandbox slightly.

I've kept the ServerSocket code, because I want to add some form of RPC
and communication between separate instances in the future.  However,
I don't expect this to be handled outside the main process, so I've
removed the Capsicum-specific connectat/bindat code.