loader: refcount LoaderBuffer pages
One less unsafe/error prone construct.  Refcounting the page as seq is
unfortunate, but still miles better than zero-filling a non-refcounted
array.  (Plus it works better for base64 decoding.)

The len field is still necessary, because old runtime doesn't support
setLenUninit.  Oh well, it's one wasted word, not the end of the world.

As for the chunks, it looks like the allocator still uses small ones for
the seq, so we're good.