javascript: type erase dfin
It works in refc.
It won't work in ARC/ORC, but that's a problem for another time.

(FWIW, I think ARC/ORC could be made work by simply adding pointers to
the destructors.  But it's fully incompatible with refc, and I won't
maintain this so long as not even ARC works with Chawan.)