cascade: basic CSS variable support
and once again, there was light...

Well, it barely works, but it's enough to get colors back on most sites.
Nested variables aren't supported for now, and shorthand expansion only
"works" for `background' (it's a hack).

Props to the W3C for introducing the C preprocessor to CSS - I was
starting to confuse it with JSSS after calc().