pager, term: replace execCmd, File
pager: I want to change mailcap command parsing, so a custom
implementation is needed.

term: now it uses a custom buffer for the output stream too.
I don't think there is much difference in performance, but stdio was in
the way.
(Also, this way it will be easier to make it async in the future.)