cssvalues: quick hack to make simple calc() expressions work
Most calc calls I've seen look something like "calc(100% - 2em)",
i.e. just a percentage value offset by a pixel value.

As it happens, we also had 24 bits available in CSSLength objects
because of padding.  Just enough to smuggle in that offset :)

Very limited of course, but it does fix layout on a couple of websites.