buffer: improve screen state synchronization
The intent behind the original design was that the page shouldn't change
without user interaction.  This sounds good in theory, but in practice
it's unnecessary if you disable JS (nothing will modify the page), and
annoying if you enable it (breaks pages that load with AJAX).

So now we queue a task to notify us when the page has been updated, and
act accordingly when the task is completed.