javascript: fix .jsget on JSValue
Since toJS doesn't create a new value, `node' must be dup'd if it is
a JSValue.

Also, forbid .jsset on JSValue as it was broken as well and it's not
obvious what the intention is when using it.  (Probably it should free
the old value, but did the user think of setting it? :P)

I can see why I made toJS not dup the value - to make returning
consistent with Nim's semantics - but it really makes things
confusing... implicitly behaving as lent might be better, albeit less
convenient.