dynstream: remove exceptions
Now we just pass down the value of n and check errno, plus
readDataLoop/writeDataLoop returns a bool indicating whether it failed.
For now this seems to work OK, but maybe I'll add a better abstraction
in the future.

EOFError is still used for handling failed packets; this is brittle,
and should be replaced once we have a proper buffering mechanism for
them.  (That will also let us kill BufStream.)

Unrelated: this also fixes a bug in buffer with cacheId.