layout: separate out tree construction logic
For now, the skeleton remains in layout.  Eventually it should be
lazily constructed during the actual layout pass (thereby making layout
"single-pass" (sometimes :p))

The end goal is to do it all in styledNode.children, so that caching can
be as simple as "next box = find next matching cached box ?? make new".

This does mean that the internal structure of cached boxes will always
have to be reconstructed, but I don't see a better way.  (I suppose it
still remains possible to optimize out the unnecessary layout pass when
only a repaint is needed (e.g. color change) by modifying computed
values in-place.)