layout: position absolute boxes relative to their parent
Also fixes an invisible bug where inline-block child absolutes were
queued multiple times.

This adds a pointer to the parent box for CSSBox objects, which isn't
great, but the alternatives (maintaining an explicit stack or adding
another tree traversal) were overly complex and/or too inefficient.

On the flip side, now it should be possible to do both stacking contexts
(with negative z-index) and overflow tracking in layout. (I think.)