render, pager: fix canvas background painting order
We were painting the background box in render for dump mode, but this
conflicted with the standard requirement that the canvas be painted
before other elements.  So now we handle this directly in the pager.

Conveniently enough, this also fixes the issue of canvas color adding
pointless spacing to pages (which often made the selection feature less
useful.)