loader: refactor handle registration, fix a cache bug
Previously, the code just called put and unset on handles in situations
where it seemed necessary, but this has become unmanageable.
Now, put is called as soon as a handle receives its stream, and unset is
called whenever said handle (and thereby its stream) is closed.

This fixes a bug in loadCGI with an rbtCache body.