loader, forkserver: launch CGI processes from fork server
Well, it's not much prettier...  but it will be useful for keep-alive.
(Hopefully.)

Also, loader could now be locked down capsicum/pledge/seccomp, but I'm
not sure if there's a point.  The biggest potential threat it faces is
cross-contamination of handles, and no amount of syscall filtering is
going to protect against that.