newhttp: clean up state machine, fix a brotli bug
The issue with brotli was that we were calling toOpenArray with a
uint, and that doesn't really work in Nim (even though it looks like
it does).

So I've removed the circular buffer from brotli and made oqoff an int
for tinfl (which unlike brotli actually needs a circular buffer).