render: fix formatting generated for padding
The old logic was badly copy pasted from the pass that did text
formatting; accordingly, it was overly complex and did not work
correctly.

Now we just add a single formatting cell to override the last one if
it exists.  (Formatting cells go from x position -> next cell or line
end, and padding implies that the line had ended before the starting
position of our text.)

I've left in a workaround to what I think is a bug.  It may be related
to cursorNextLink getting stuck at the line's end...