config.toml: restore old UA string
Originally I added Mozilla/5.0 because Google refused to serve valid
HTML without it, and some websites really disliked us not sending it.
Since then, I've had to special case Google anyway because now it only
serves usable results to a specific set of UA strings.

Meanwhile, an aggressive client blocking software has been spreading
lately, which shows a (broken, for various reasons) JS POW page if you
*do* send Mozilla/5.0.  So I guess it's better to switch back for now.

The drawback of course is that now we are broken on sites that expect
us to send Mozilla/5.0.  Hopefully this is no longer as common as
before.