loader: use pipes in network streams
Turns out we don't actually have a reliable mechanism to detect broken
streams - if a pipe is broken, well, it's broken, but that's about all
we know.  It seems UNIX sockets didn't trigger the code path, but pipes
do now, and it was causing issues (reporting some correctly ended
streams as broken).

This is not fixable with the current protocol, so for now, I've just
made onError do the same as onRead at EOF.