term: replace termcap with a built-in terminal database
I considered using terminfo, but it works *worse* on several terminals
I've tried than whatever this is.  So out with ncurses, I say.

In this new scheme, if the terminal is unrecognized, it is assumed to
be XTerm-compatible.  This does mean that an unrecognized terminal that
doesn't do XTGETTCAP or COLORTERM will not get true colors.

Another drawback is that non-VT-compatible terminals now have zero
chance of working.  This is unfortunate, but realistically they probably
didn't work until now either.