dom: improve script processing
We now only mark classic scripts as ready after their onReady is set.
This way, we can avoid the event loop issue where the spec assumes that
the fetch will only finish after the function has returned.

(I believe it is be possible to make prepare resumable, but it would
require nested invocation of the event loop.)