layout: performance hack for table layout
This breaks max-content rendering (which isn't really exposed to
websites anyway).  In return, it greatly increases performance on
sites that rely on nested tables for layout.

A better solution would be to skip the second layout during the
measurement phase of max-content and fit-content descendants, but that
will require a significant refactoring of the entire module first.