opt: add := operator
The lack of a `var' or `let' is ugly, but it's still cleaner (and more
efficient) than assigning a temporary.

An alternate syntax I've considered is

if (var x: T; foo().into x):
  bar(x)

but this way, a) I can't use type interference for x, b) I have to
declare x before the funcall which makes it harder for the compiler to
optimze out the zero init.