Improve CSS variable support
Now it works according to the spec, for the most part.  I suspect it's
not 100% right in edge cases where a component value bleeds into
another, but whatever.  At least now we render Codeberg correctly.

This required deforestation of CSSFunction.  Other tree-style tokens
remain because I don't want to risk more regressions than necessary.
Of course, this doesn't make the parser any less incomprehensible.

Also, we now do somewhat more copies than before; I think it's a
performance loss overall, but it doesn't seem to have significant
effects in practice.  It's not so bad if you consider this a first step
towards unifying cssvalues and cssparser (which will reduce the number
of CSS parser passes).