loader: fix loading of partially cached handles
Fixes regression in 20bc1755a.  The old code used deques, and with
that parent switching worked fine.  However, now that new buffers are
just attached to the parent handle's chain, the old parent's chain
must be connected with that of the new one.

The effect was that sometimes, images would not fully load on the first
try, only when loaded from the cache.