md2html: parse pre-like tags in inline
We try to imitate pandoc instead of CommonMark in parsing HTML, which
means you can format HTML block contents, but also that e.g. <style>
inside an HTML head tag doesn't work as expected.

Until now, we just treated head as pre-like, but that stops working as
soon as you put an <html> in the document too.  So now we also parse
pre-like tags in inline too, which has the benefit that you also get to
put <style> or <pre> anywhere (mostly).