Fix compilation with ORC
Crucially, it still doesn't *work*, because in ORC, casting non-RootRef
pointers to RootRef is not supported.  The easiest way out is to make
all registered types RootRef, but that would be a waste of memory as
long as we are still on refc.

For now, the primary goal is just to make comparing generated object
size easier.  (I guess it also makes Monoucha work on ORC with some
caveats, that should be documented.)