pager, term: rewrite input handler, add bracketed paste
* unblock stdin on startup
* decode input as the correct charset
* add bracketed paste mode

Unblocked stdin has two advantages.  First, it paves way for a future
rewrite of queryAttrs which could work around GNU screen's refusal to
print the foreground color & DA1 in the right order.  Second, Linux's
documentation notes a bug where a poll may return a read for a blocking
file descriptor, so that polling a blocking stdin is in fact plain
broken.

Bracketed paste allows for more sensible handling of paste events:
if any text is pasted to the pager, it is automatically loaded to the
URL bar.  This is superior to M-p in two aspects: a) xsel is not needed
(i.e. it works over SSH), b) it allows the user to edit the URL before
opening it.

This also fixes some miscellaneous issues in the previous
implementation, which had spaghettified beyond comprehension.