config: remove [protocol] map
The original idea was to eliminate any kind of protocol-dependency from
the binary, but putting protocol definitions in different places isn't
the right way to do this.

Instead, we should deprecate urimethodmap in favor of a mailcap-like
config file for protocol definitions, a la w3mmee.  (But even then, it's
questionable whether it's worth specifying submission methods instead of
just converting the URL in CGI.)