func -> proc
`func` gives the compiler too little information to optimize anything
(hence experimental:strictFuncs), enables no new capabilities like
async/await, and the safety it provides is also dubious (a func
can still modify some completely unrelated object through five ref
indirections.)

It helps somewhat that the compiler tries to infer noSideEffects, but
the inference breaks down as soon as you forward declare a proc.
Overall it's worse than useless, so kill it with fire.

(I've left it in for chame for now, because it's the only place where
breaking the API may be problematic.)