pager: add click & drag mouse selection
This was hard to squeeze into the existing UI.

First I tried so it would always select on click & drag.  But after playing
around with it for a bit, I was getting annoyed because it kept selecting
text when I was just trying to move the screen.

So then I tried to do it with triple click, but that in turn didn't work
on links.  Also I had to click around so much it was starting to get
ridiculous.

In the end I've settled on this system: click + drag on the same row
starts a selection, while click + drag to a different row moves the screen.
It's not perfect, because now selection is somewhat unintuitive, but I think
it's the best I can do with w3m-style mouse handling.