layout: basic table border support
It still only works with one border style, and doesn't handle edge cases
such as tables with uneven row sizes too well.  But the fundamentals seem
right.

I'm also not sure yet what to do with border-spacing; for now, it's ignored
as soon as the table has any border, but that feels very arbitrary (and
doesn't really work if only cells have borders).