Saner error handling in pager
atexit is not specified to handle the things we were throwing at it, and the
cleanup logic was flawed anyway, e.g. if a JS function triggered a write on
closed stdout then we'd try to free the runtime inside a JS call...

So now we just propagate errors and call exit once at the program's end.
Note that mailcap processes ignore these; I guess this should be changed
eventually, but it doesn't seem too important (the next draw will fail
anyway, and then we quit.)

Also:
* move die() back to main, as it really shouldn't be called after init or in
  other processes
* remove the undocumented flushConsole JS function which has lost its
  purpose