javascript: remove support for exposing value types
This was less useful than it seems: in cases where memory matters, it's more
effective to lazy-initialize a ref object (either by defaulting to nil as an
internal slot, or using a WeakMap).  As a result it was only used in cases
where typically there is only one object in the entire lifetime of the
program, and thus barely tested.

I guess for cases such as DOMImplementation it would be nice to support
multiple JSValues for one Nim ref, but I really don't feel like messing
around with lifetimes anymore.