config.toml: replace broken Google search with Brave search
They killed all ways to access without JS.  If I do enable JS as instructed,
it returns a captcha after a long series of redirects.  Aren't monopolies
great?

I've switched to Brave for now as it seems to return the least awful results
of the alternatives.  It has no colors for some reason but the layout looks
better than I remember so I guess it's fine.