layout: simplify/refactor margin collapsing logic
Compared to the previous version, this should be more efficient (less
allocations), as well as (hopefully) more correct.  More importantly, it
no longer depends on state global to the entire BFC, bringing us closer to
being able to cache block layout.

(The Acid1 change seems to be correct; the box to the right has a shorter
height than that on the left, and we do not (cannot) compensate for the
difference with borders.)