csstree, layout: cache box tree between layouts
Took a lot of refactoring, but it's finally time.

As expected, it does wonders; on larger pages it reduces relayout time by
orders of magnitude.  Even on smaller pages the memory usage is much lower,
since we no longer allocate a whole DOM of boxes for each relayout.

There is still room for improvement (what with the two tree traversals),
but this is the point where I'm finally satisfied with performance.

Also fixed some bugs that have become more visible with caching.