pager: render previous buffer if current one is empty
This is similar to what pretty much all other browsers do.  Combined with
the previous commit it means that page transitions are now smooth, without
any flashing (unless the page itself is doing something stupid).

The drawback is that it's no longer obvious what buffer you're looking at;
to counter this, trying to interact with the buffer (move cursor/click)
swaps to the real buffer which should result in halfway intuitive behavior.