dirlist2html: eliminate wcwidth call
Apparently the binding crashed even on Linux under some libcs.  And of
course some systems use 16-bit wcwidth, etc...

So instead, we now use a CSS hack that *almost* displays the same results.
In some cases it does show a bit larger border than desired, but I guess
that's the price of progress.