pager, buffer: fix unregistered handle skipping logic
If a handle is already unregistered in the current cycle, then we want to
skip it no matter if another handle has taken its place.

While the previous behavior was incorrect, it mattered very rarely; I only
noticed it because I was getting a "loader died" instead of a "no CGI file"
error with an incorrectly configured CGI dir.

In fact, for anything but error the previous behavior was *technically*
fine, as the worst you can get is an EAGAIN or a blocking read.  In practice
we really don't want the pager to block, so it had to be fixed too.