pager: actually fix x-ansioutput
I'm kind of conflicted about this feature, the implementation is a kludge
that depends on another kludge (x-needsstyle), and it never worked correctly
anyway, so I was on the verge of removing it.

On the other hand, there's no good alternative for handling programs that
print ANSI-formatted text.  And the dependency on x-needsstyle can be
removed later by properly switching the content type to text/x-ansi.