buffer: store prevHover in pager handle
Storing it in BufferContext had the issue that clicking on an anchor on
the same page would clear the previous hover, which looks awkward when you
go back.

(It's still a bit quirky since the two handles don't sync their hover
status.  An alternative to avoid this is to send updateHover on buffer
switch, but that makes it observable by JS which I really want to avoid...)