term: rework terminal query handling
The previous solution (send queries & blocking read responses) had some
issues:

* we had to wait for the terminal before starting any network requests,
  which is obviously sub-optimal
* if user input came before terminal input, we got very confused (this is
  rare, but it can happen on slow connections)
* some terminals don't respond to DA1 so it barely worked
* GNU screen responds to foreground/background queries out of order, so we
  couldn't send the queries at all (even on tmux which advertises itself
  as screen)

Now we just let query responses interleave with user input.  If we get a
response that changes rendering, we redraw or windowChange as appropriate.

Also, instead of using DA1 as a sentinel, we use the good old VT100 resize
hack.  Whether or not this is better emulated than the DA1 remains to be
seen.  (I suspect it is, but at this point it doesn't really matter...)